Finalize rebuild
