+++ lib/repositories/auth_repository.dart
@@ -1,6 +1,8 @@
 import 'package:allo/interface/home/stack_navigator.dart';
-import 'package:allo/interface/login/signup.dart';
-import 'package:allo/interface/login/login.dart';
+import 'package:allo/interface/login/existing/enter_password.dart';
+import 'package:allo/interface/login/new/setup_name.dart';
+import 'package:allo/interface/login/new/setup_password.dart';
+import 'package:allo/interface/login/new/setup_verification.dart';
 import 'package:allo/main.dart';
 import 'package:allo/repositories/preferences_repository.dart';
 // import 'package:allo/repositories/preferences_repository.dart';
@@ -10,7 +12,10 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:image_cropper/image_cropper.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 // Spec:
@@ -94,33 +99,93 @@ class AuthRepository {
   Future signUp(
       {required String email,
       required String password,
+      required String confirmPassword,
       required String displayName,
       required String username,
       required ValueNotifier<String> error,
       required BuildContext context}) async {
     try {
-      final user = await FirebaseAuth.instance
-          .createUserWithEmailAndPassword(email: email, password: password);
-      await context.read(preferencesProvider).setBool('isAuth', true);
-      await user.user!.updateDisplayName(displayName);
-      final db = FirebaseFirestore.instance;
-      await db.collection('users').doc(username).set({
-        'name': displayName,
-        'email': email,
-        'uid': user.user!.uid,
-        'verified': false,
-      });
-      await db.collection('users').doc('usernames').update({
-        username: user.user!.uid,
-      });
-      await ref
-          .read(navigationProvider)
-          .push(context, StackNavigator(), SharedAxisTransitionType.horizontal);
+      FocusScope.of(context).unfocus();
+      if (password != '' && confirmPassword != '') {
+        if (password == confirmPassword) {
+          var hasUppercase = password.contains(RegExp(r'[A-Z]'));
+          var hasDigits = password.contains(RegExp(r'[0-9]'));
+          var hasLowercase = password.contains(RegExp(r'[a-z]'));
+          var hasSpecialCharacters =
+              password.contains(RegExp(r'[!@#$%^&*(),.?":{}|<>]'));
+          var hasMinLength = password.length >= 8;
+          if (hasUppercase &&
+              hasDigits &&
+              hasLowercase &&
+              hasSpecialCharacters &&
+              hasMinLength) {
+            final user = await FirebaseAuth.instance
+                .createUserWithEmailAndPassword(
+                    email: email, password: password);
+            await context.read(preferencesProvider).setBool('isAuth', true);
+            await user.user!.updateDisplayName(displayName);
+            final db = FirebaseFirestore.instance;
+            await db.collection('users').doc(username).set({
+              'name': displayName,
+              'email': email,
+              'uid': user.user!.uid,
+              'verified': false,
+            });
+            await db.collection('users').doc('usernames').update({
+              username: user.user!.uid,
+            });
+            await ref.read(navigationProvider).push(context,
+                SetupVerification(), SharedAxisTransitionType.horizontal);
+          } else {
+            error.value = 'Parola ta nu respectă cerințele.';
+          }
+        } else {
+          error.value = 'Parolele nu sunt la fel.';
+        }
+      } else {
+        error.value = 'Parolele nu trebuie să fie goale.';
+      }
     } on FirebaseAuthException catch (e) {
       switch (e.code) {
         case 'operation-not-allowed':
           error.value = 'Momentan, înregistrările sunt închise.';
       }
+    }
+  }
+
+  Future checkUsernameInSignUp(
+      {required String username,
+      required ValueNotifier<String> error,
+      required BuildContext context,
+      required String displayName,
+      required String email}) async {
+    final usernameReg = RegExp(r'^[a-zA-Z0-9_\.]+$');
+    final navigation = context.read(Repositories.navigation);
+    final usernamesDoc = await FirebaseFirestore.instance
+        .collection('users')
+        .doc('usernames')
+        .get();
+    final usernames = usernamesDoc.data() as Map;
+    FocusScope.of(context).unfocus();
+    if (username != '') {
+      if (usernameReg.hasMatch(username)) {
+        if (!usernames.containsKey(username)) {
+          await navigation.push(
+              context,
+              SetupPassword(
+                displayName: displayName,
+                username: username,
+                email: email,
+              ),
+              SharedAxisTransitionType.horizontal);
+        } else {
+          error.value = 'Acest nume de utilizator este deja luat.';
+        }
+      } else {
+        error.value = 'Numele de utilizator nu este valid.';
+      }
+    } else {
+      error.value = 'Numele de utilizator nu poate fi gol.';
     }
   }
 
@@ -230,6 +295,50 @@ class AuthRepository {
       final name = userDataMap['name'];
       await prefs.setString('displayName', name);
     }
+  }
+
+  Future updateProfilePicture(
+      {required ValueNotifier<bool> loaded,
+      required ValueNotifier<double> percentage,
+      required BuildContext context,
+      Widget? route}) async {
+    XFile imageFile;
+    var pickFromGallery =
+        await ImagePicker().pickImage(source: ImageSource.gallery);
+    var uneditedImageFile = XFile(pickFromGallery!.path);
+    if (kIsWeb) {
+      imageFile = uneditedImageFile;
+      loaded.value = true;
+    } else {
+      var editImageFile = await ImageCropper.cropImage(
+          sourcePath: pickFromGallery.path,
+          aspectRatioPresets: [CropAspectRatioPreset.square]);
+      var convertedEditImageFile = XFile(editImageFile!.path);
+      imageFile = convertedEditImageFile;
+      loaded.value = true;
+    }
+    var user = FirebaseAuth.instance.currentUser;
+    var filePath = 'profilePictures/${user?.uid}.png';
+    var uploadTask = FirebaseStorage.instance.ref(filePath).putData(
+        await imageFile.readAsBytes(),
+        SettableMetadata(contentType: 'image/png'));
+    uploadTask.snapshotEvents.listen((TaskSnapshot snapshot) async {
+      percentage.value =
+          (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
+      if (snapshot.state == TaskState.success) {
+        await user!.updatePhotoURL(await FirebaseStorage.instance
+            .ref()
+            .child(filePath)
+            .getDownloadURL());
+        if (route == null) {
+          Navigator.pop(context);
+        } else {
+          await context
+              .read(navigationProvider)
+              .push(context, route, SharedAxisTransitionType.horizontal);
+        }
+      }
+    });
   }
 }
